# tacocloud/tacocloud-api/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Taco Cloud API
  version: 1.0.0
  description: >-
    Public v1 contract. Legacy /api routes remain available until 2027-12-31 and return
    Deprecation, Sunset and Link headers pointing to the equivalent /api/v1 route.
servers:
  - url: /api/v1
...
paths:
...
  /tacos:
    get:
      tags: [Tacos]
      operationId: listTacos
      parameters:
        - {name: name, in: query, schema: {type: string}}
        - {name: ingredientId, in: query, schema: {type: string}}
        - {name: diet, in: query, schema: {type: string}}
        - {name: excludeAllergen, in: query, schema: {type: string}}
        - {name: spice, in: query, schema: {type: string}}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Size'}
        - {name: sort, in: query, schema: {type: string, default: 'createdAt,desc'}}
        - {name: recent, in: query, schema: {type: string}}
      responses:
        '200': {$ref: '#/components/responses/TacoPage'}
        '400': {$ref: '#/components/responses/Problem'}
    post:
      tags: [Tacos]
      operationId: createTaco
      security: [{cookieAuth: []}]
      requestBody: {$ref: '#/components/requestBodies/Taco'}
      responses:
        '201': {$ref: '#/components/responses/Taco'}
        '400': {$ref: '#/components/responses/Problem'}
        '422': {$ref: '#/components/responses/Problem'}
...
  /tacos/validate:
    post:
      tags: [Tacos]
      operationId: validateTaco
      requestBody: {$ref: '#/components/requestBodies/Taco'}
      responses:
        '200': {$ref: '#/components/responses/TacoValidation'}
        '400': {$ref: '#/components/responses/Problem'}
...
  /tacos/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Tacos]
      operationId: getTaco
      responses:
        '200': {$ref: '#/components/responses/Taco'}
        '404': {$ref: '#/components/responses/Problem'}
  /tacos/{id}/classification:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Tacos]
      operationId: classifyTaco
      responses:
        '200': {$ref: '#/components/responses/TacoClassification'}
        '404': {$ref: '#/components/responses/Problem'}
...
  /orders:
    post:
      tags: [Orders]
      operationId: createOrder
      security: [{cookieAuth: []}]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OrderCreateRequest'}
            example:
              deliveryName: Daniel
              deliveryStreet: Calle 1
              deliveryCity: Guadalajara
              deliveryState: Jalisco
              deliveryZip: '44100'
              paymentMethodId: pm_123
              items:
                - quantity: 2
                  taco:
                    name: Taco especial
                    ingredientIds: [FLTO, GRBF]
      responses:
        '201':
          description: Created order or the original response for an idempotent retry
          headers:
            X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OrderResponse'}
        '400': {$ref: '#/components/responses/Problem'}
        '401': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
        '422': {$ref: '#/components/responses/Problem'}
...
  /orders/{orderId}:
    parameters:
      - {name: orderId, in: path, required: true, schema: {type: string}}
    patch:
      tags: [Orders]
      operationId: patchOrder
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OrderPatchRequest'}
      responses:
        '200': {$ref: '#/components/responses/Order'}
        '400': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
    put:
      tags: [Orders]
      operationId: replaceOrder
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OrderReplaceRequest'}
      responses:
        '200': {$ref: '#/components/responses/Order'}
        '400': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
    delete:
      tags: [Orders]
      operationId: deleteOrder
      security: [{cookieAuth: []}]
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '404': {$ref: '#/components/responses/Problem'}
  /orders/{id}/status:
    parameters:
      - $ref: '#/components/parameters/Id'
    patch:
      tags: [Orders]
      operationId: updateOrderStatus
      security: [{cookieAuth: []}]
      requestBody: {$ref: '#/components/requestBodies/OrderStatus'}
      responses:
        '200': {$ref: '#/components/responses/OrderDetail'}
        '400': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
  /orders/{id}/cancel:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [Orders]
      operationId: cancelOrder
      security: [{cookieAuth: []}]
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OrderCancellationRequest'}
      responses:
        '200': {$ref: '#/components/responses/OrderDetail'}
        '409': {$ref: '#/components/responses/Problem'}
...
  /users/me/orders:
    get:
      tags: [Orders]
      operationId: listMyOrders
      security: [{cookieAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Size'}
      responses:
        '200': {$ref: '#/components/responses/OrderPage'}
  /users/me/orders/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Orders]
      operationId: getMyOrder
      security: [{cookieAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/OrderDetail'}
        '404': {$ref: '#/components/responses/Problem'}
...
  /kitchen/orders/{id}/status:
    parameters:
      - $ref: '#/components/parameters/Id'
    patch:
      tags: [Kitchen]
      operationId: updateKitchenOrderStatus
      security: [{cookieAuth: []}]
      requestBody: {$ref: '#/components/requestBodies/OrderStatus'}
      responses:
        '200': {$ref: '#/components/responses/KitchenOrder'}
        '409': {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: {type: string}
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 0, default: 0}
    Size:
      name: size
      in: query
      schema: {type: integer, minimum: 1, maximum: 50, default: 20}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 8, maxLength: 100, pattern: '^[A-Za-z0-9._:-]+$'}
  headers:
    CorrelationId:
      description: Correlation identifier accepted from the request or generated by the server
      schema: {type: string}
  requestBodies:
...
    Taco:
      required: true
      content:
        application/json:
          schema: {$ref: '#/components/schemas/TacoCreateRequest'}
    OrderStatus:
      required: true
      content:
        application/json:
          schema: {$ref: '#/components/schemas/OrderStatusUpdateRequest'}
  responses:
    NoContent:
      description: Operation completed without a response body
      headers:
        X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
    Problem:
      description: RFC 9457 compatible API error
      headers:
        X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/ApiProblem'}
          example:
            type: urn:tacocloud:problem:validation-error
            title: Request validation failed
            status: 400
            detail: One or more fields are invalid
            instance: /api/v1/orders
            code: VALIDATION_ERROR
            violations: [{field: deliveryZip, reason: must contain exactly 5 digits}]
...
    Taco:
      description: Taco response
      content: {application/json: {schema: {$ref: '#/components/schemas/TacoResponse'}}}
    TacoPage:
      description: Taco page or recent taco list
      content: {application/json: {schema: {$ref: '#/components/schemas/TacoPageResponse'}}}
...
    TacoValidation:
      description: Taco validation result
      content: {application/json: {schema: {$ref: '#/components/schemas/TacoValidationResponse'}}}
    TacoClassification:
      description: Taco classification
      content: {application/json: {schema: {$ref: '#/components/schemas/TacoClassificationResponse'}}}
...
    Order:
      description: Order response
      content: {application/json: {schema: {$ref: '#/components/schemas/OrderResponse'}}}
    OrderDetail:
      description: Detailed order response
      content: {application/json: {schema: {$ref: '#/components/schemas/OrderDetailResponse'}}}
    OrderPage:
      description: Order history page
      content: {application/json: {schema: {$ref: '#/components/schemas/OrderPageResponse'}}}
...
    KitchenOrder:
      description: Kitchen order
      content: {application/json: {schema: {$ref: '#/components/schemas/KitchenOrderResponse'}}}
...
  schemas:
...
    IngredientResponse:
      type: object
      required: [id, name, type, unitPrice, available]
      properties:
        id: {type: string}
        name: {type: string}
        type: {$ref: '#/components/schemas/IngredientType'}
        unitPrice: {type: number}
        available: {type: boolean}
        dietaryTags: {type: array, items: {type: string}}
        allergens: {type: array, items: {type: string}}
        spiceLevel: {type: string}
    IngredientType:
      type: string
      enum: [WRAP, PROTEIN, VEGGIES, CHEESE, SAUCE]
    TacoCreateRequest:
      type: object
      additionalProperties: false
      required: [name, ingredientIds]
      properties:
        name: {type: string, minLength: 5, maxLength: 50}
        ingredientIds: {type: array, minItems: 1, maxItems: 10, items: {type: string}}
    TacoResponse:
      type: object
      required: [id, name, ingredients]
      properties:
        id: {type: string}
        name: {type: string}
        createdAt: {type: string, format: date-time}
        ingredients: {type: array, items: {$ref: '#/components/schemas/IngredientResponse'}}
        classification: {$ref: '#/components/schemas/TacoClassificationResponse'}
    TacoPageResponse:
      type: object
      properties:
        content: {type: array, items: {$ref: '#/components/schemas/TacoResponse'}}
        page: {type: integer}
        size: {type: integer}
        totalElements: {type: integer, format: int64}
        totalPages: {type: integer}
        sort: {type: string}
...
    TacoValidationResponse:
      type: object
      properties:
        valid: {type: boolean}
        violations: {type: array, items: {$ref: '#/components/schemas/ApiViolation'}}
    TacoClassificationResponse:
      type: object
      properties:
        dietaryTags: {type: array, items: {type: string}}
        allergens: {type: array, items: {type: string}}
        spiceLevel: {type: string}
...
    OrderTacoRequest:
      type: object
      additionalProperties: false
      required: [name, ingredientIds]
      properties:
        name: {type: string, minLength: 5, maxLength: 50}
        ingredientIds: {type: array, minItems: 1, maxItems: 10, items: {type: string}}
    OrderItemRequest:
      type: object
      additionalProperties: false
      required: [taco, quantity]
      properties:
        taco: {$ref: '#/components/schemas/OrderTacoRequest'}
        quantity: {type: integer, minimum: 1}
    OrderCreateRequest:
      type: object
      additionalProperties: false
      required: [deliveryName, deliveryStreet, deliveryCity, deliveryState, deliveryZip, paymentMethodId, items]
      properties:
        deliveryName: {type: string, maxLength: 100}
        deliveryStreet: {type: string, maxLength: 120}
        deliveryCity: {type: string, maxLength: 80}
        deliveryState: {type: string, maxLength: 80}
        deliveryZip: {type: string, pattern: '^\d{5}$'}
        paymentMethodId: {type: string}
        couponCode: {type: string, maxLength: 40}
        items: {type: array, minItems: 1, maxItems: 20, items: {$ref: '#/components/schemas/OrderItemRequest'}}
    OrderResponse:
      type: object
      required: [id, status, userId, deliveryName, deliveryStreet, deliveryCity, deliveryState, deliveryZip, subtotal, discount, total, currency, items]
      properties:
        id: {type: string}
        placedAt: {type: string, format: date-time}
        status: {$ref: '#/components/schemas/OrderStatus'}
        userId: {type: string}
        deliveryName: {type: string}
        deliveryStreet: {type: string}
        deliveryCity: {type: string}
        deliveryState: {type: string}
        deliveryZip: {type: string}
        paymentMethod: {$ref: '#/components/schemas/PaymentMethodResponse'}
        subtotal: {type: number}
        discount: {type: number}
        total: {type: number}
        currency: {type: string, example: MXN}
        couponCode: {type: string, nullable: true}
        items: {type: array, items: {$ref: '#/components/schemas/OrderItemResponse'}}
    OrderItemResponse:
      type: object
      properties:
        taco: {$ref: '#/components/schemas/TacoResponse'}
        quantity: {type: integer}
        unitPrice: {type: number}
        lineTotal: {type: number}
    OrderPatchRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        deliveryName: {type: string, maxLength: 100}
        deliveryStreet: {type: string, maxLength: 120}
        deliveryCity: {type: string, maxLength: 80}
        deliveryState: {type: string, maxLength: 80}
        deliveryZip: {type: string, pattern: '^\d{5}$'}
    OrderReplaceRequest:
      type: object
      additionalProperties: false
      required: [deliveryName, deliveryStreet, deliveryCity, deliveryState, deliveryZip, tacos]
      properties:
        deliveryName: {type: string}
        deliveryStreet: {type: string}
        deliveryCity: {type: string}
        deliveryState: {type: string}
        deliveryZip: {type: string, pattern: '^\d{5}$'}
        tacos: {type: array, minItems: 1, maxItems: 20, items: {$ref: '#/components/schemas/OrderReplacementTacoRequest'}}
    OrderReplacementTacoRequest:
      type: object
      required: [name, ingredients]
      properties:
        name: {type: string, minLength: 5}
        ingredients: {type: array, minItems: 1, items: {$ref: '#/components/schemas/IngredientResponse'}}
    OrderStatus:
      type: string
      enum: [CREATED, PLACED, ACCEPTED, PREPARING, READY, OUT_FOR_DELIVERY, DELIVERED, CANCELLED]
    OrderStatusUpdateRequest:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status: {$ref: '#/components/schemas/OrderStatus'}
        reason: {type: string, maxLength: 200}
    OrderCancellationRequest:
      type: object
      additionalProperties: false
      properties:
        reason: {type: string, maxLength: 200}
    OrderDetailResponse:
      allOf:
        - $ref: '#/components/schemas/OrderResponse'
        - type: object
          properties:
            statusHistory: {type: array, items: {type: object}}
    OrderPageResponse:
      type: object
      properties:
        content: {type: array, items: {$ref: '#/components/schemas/OrderResponse'}}
        page: {type: integer}
        size: {type: integer}
        totalElements: {type: integer, format: int64}
        totalPages: {type: integer}
...
    KitchenOrderResponse:
      type: object
      properties:
        id: {type: string}
        placedAt: {type: string, format: date-time}
        status: {$ref: '#/components/schemas/OrderStatus'}
        stationId: {type: string}
        cookId: {type: string, nullable: true}
        acceptedAt: {type: string, format: date-time, nullable: true}
        estimatedPrepMinutes: {type: integer}
        items: {type: array, items: {type: object}}
...
    PaymentMethodResponse:
      type: object
      properties:
        paymentMethodId: {type: string}
        brand: {type: string}
        last4: {type: string, pattern: '^\d{4}$'}
...
    ApiProblem:
      type: object
      required: [type, title, status, detail, instance, code, violations]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        detail: {type: string}
        instance: {type: string, format: uri-reference}
        code: {type: string}
        violations: {type: array, items: {$ref: '#/components/schemas/ApiViolation'}}
    ApiViolation:
      type: object
      required: [field, reason]
      properties:
        field: {type: string}
        reason: {type: string}